requirements.txt
streamlit
google-generativeai
pinecone-client
sentence-transformers